+++ jdcrafte/CHANGELOG.md
@@ -6,6 +6,19 @@
   wheat seeds) is added. Right-click with feed in hand to fill it, or shift-right-click with an
   empty hand to take feed back out.
 * Added the weathervane, a decorative rooftop weather vane. Right-click to spin it.
+* Added the Home Sweet Hrmmm plaque, a two-block-wide decorative wall sign. Comes in every wood type:
+  * Acacia
+  * Bamboo
+  * Birch
+  * Cherry
+  * Crimson
+  * Dark Oak
+  * Jungle
+  * Mangrove
+  * Oak
+  * Pale Oak
+  * Spruce
+  * Warped
 * Added the trellis, a climbable decorative wooden lattice, as well as trellis arch. Comes in every wood type:
   * Acacia
   * Bamboo